Fix bug in moving common Functions in evaluation
The Functions were removed from ALL parents, including remote parents,
and added to the root as Functions. This resulted in adding them TWICE
to the root if there was a remote parent, or more if there were multiple
remote parents.
Now remote parents are preserved, and the unit is only removed from the
primary parent and added to the root.