Remove &quot; characters from test extra remarks
Looks like from_standard strips the quotes. It didn't fail before
because it just wasn't called in the test.
It strips the quotes because it tries to read the extra entries as JSON,
and if it succeeds, it returns the decoded JSON. Otherwise it returns
the original string. Just remarks is not valid JSON, but "remarks" is,
and is decoded to the string remarks... I'd like to keep this JSON
decoding capability (since sometimes we save actual dicts of values to
extra entries), so I just removed the extra &quot;s from standard3.xml.

This means that reading and writing an XML is not promised to return the
same XML: if there are bare strings in there, they will always be
written without quotes even if they originally had them. An alternative
would be to mark in _loads() in convert.py whether the original string
was a JSON value or just a string, but I think this would be an
overkill. No one writes quoted strings in the comments anyway, so it's
not that important to test it.

See danielhers/ucca#56.