Upload full tokens in children_tokens in json
We changed the API slightly, which gives me now issues with the passages you uploaded. The change is that in the 'children_tokens' field of annotation units we now put the full token (the same object you put in the 'tokens' field in the task object) and not only the token id.